add theme blog

+++ app/Http/Controllers/Categorycontroller.php
@@ -92,6 +92,10 @@ class Categorycontroller extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Categoria nao pode ser deletada porque tem posts');
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Category deletede successfully');
